feat(mic): added mic mute action
